Create CTS-Code
Hi there! Thanks for taking the time. This is a small fun little project that I started about a year ago but I can't seem to get it right. I have spent to many hours trying to figure out whats wrong and make the code more readable. I will update and add coments so you can understand what part of the code does what more easily. Will have to freshen up on the code before doing this. 
The Problem:
Say for example in 1000 tournaments the 2nd rated winning 88%, the 1st rated winning 12% of the time and no one else with even one tournament victory. Other times its 3rd which has way more wins then it should. I was expecting a descending curve but this is everything but that. 

If you want  to run the code by yourself:
Insert path of the Exel file in line 13. You probably can change it to other file formats (csv, ...) but if you do that you would have to change the extractData() function.